🎨 年の選択部分をコンポーネント化
